Added locale feature specs
Necessary to ensure we have parity with English (default) translations by looping through all files, flattening the keys, and checking for any differences. Zero differences is good because that means we have identical keys. A separate spec is supplied for category of locale despite some duplication in setup which seems acceptable for now.

In order to check key parity, the locale of each file loaded is always ignored in order to ensure equal comparisons.