Fixed raw locale file names
This was a mistake on my part. I couldn't see `raw.yml` via the `I18n.load_path` when first converting this project to a gem but am now able to confirm that it loads properly so there is no reason to use `en.raw.yml` anymore. This also adheres to what is documented in the guide.