Removed Git Lint gem
We are removing the linter in order to make community contributions easier at the cost of degrading Git commit history and release note generation. The majority of this gem houses translations with very little code (this is on purpose) that happens to be wrapped as a gem.

This is a partial revert of 90ba60b68ffb (Added project skeleton) when first setting up the project.

Milestone: minor